requirements.txt
streamlit
snowflake-snowpark-python
pandas
snowflake
python-dotenv

